git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/gccgo@187127 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/go/syscall/libcall_posix.go
@@ -138,7 +138,7 @@ func (w WaitStatus) TrapCause() int
 //sys	Select(nfd int, r *FdSet, w *FdSet, e *FdSet, timeout *Timeval) (n int, err error)
 //select(nfd int, r *FdSet, w *FdSet, e *FdSet, timeout *Timeval) int
 
-const nfdbits = unsafe.Sizeof(fds_bits_type) * 8
+const nfdbits = int(unsafe.Sizeof(fds_bits_type) * 8)
 
 type FdSet struct {
 	Bits [(FD_SETSIZE + nfdbits - 1) / nfdbits]fds_bits_type
@@ -268,7 +268,7 @@ func Gettimeofday(tv *Timeval) (err error) {
 //sys	Mknod(path string, mode uint32, dev int) (err error)
 //mknod(path *byte, mode Mode_t, dev _dev_t) int
 
-//sys	Mount(source string, target string, fstype string, flags int, data string) (err error)
+//sys	Mount(source string, target string, fstype string, flags uintptr, data string) (err error)
 //mount(source *byte, target *byte, fstype *byte, flags _C_long, data *byte) int
 
 //sys	Nanosleep(time *Timespec, leftover *Timespec) (err error)
